+++ src/Gnugat/Redaktilo/Command/LineRemoveCommand.php
@@ -13,14 +13,15 @@ namespace Gnugat\Redaktilo\Command;
 
 use Gnugat\Redaktilo\Converter\LineContentConverter;
 
+/**
+ * Removes the given location in the given file.
+ */
 class LineRemoveCommand implements Command
 {
     /** @var LineContentConverter */
-    protected $converter;
+    private $converter;
 
-    /**
-     * @param LineContentConverter $converter
-     */
+    /** @param LineContentConverter $converter */
     public function __construct(LineContentConverter $converter)
     {
         $this->converter = $converter;
@@ -30,7 +31,7 @@ class LineRemoveCommand implements Command
     public function execute(array $input)
     {
         $file = $input['file'];
-        $location = $input['location'];
+        $location = isset($input['location']) ? $input['location'] : $file->getCurrentLineNumber();
 
         $lines = $this->converter->from($file);
         unset($lines[$location]);