+++ src/Gnugat/Redaktilo/Editor.php
@@ -158,74 +158,58 @@ class Editor
     }
 
     /**
-     * Inserts the given line before the current one.
+     * Inserts the given line before the given line number
+     * (or before the current one if none provided).
      * Note: the current line is then set to the new one.
      *
-     * @param File     $file
-     * @param string   $addition
-     * @param int|null $location
-     *
-     * @throws \Gnugat\Redaktilo\Search\NotSupportedException If the current line number isn't supported by any registered strategy
+     * @param File    $file
+     * @param string  $addition
+     * @param integer $location
      *
      * @api
      */
-    public function addBefore(File $file, $addition, $location=null)
+    public function addBefore(File $file, $addition, $location = null)
     {
-        if ($location === null || !is_integer($location)) {
-            $location = $file->getCurrentLineNumber();
-        }
-
         $input = array(
             'file' => $file,
             'location' => $location,
             'addition' => $addition,
         );
-        $this->commandInvoker->run('insert', $input);
-
-        $file->setCurrentLineNumber($location);
+        $this->commandInvoker->run('insert_above', $input);
     }
 
     /**
-     * Inserts the given line after the current one.
+     * Inserts the given addition after the given line number
+     * (or after the current one if none provided).
      * Note: the current line is then set to the new one.
      *
-     * @param File     $file
-     * @param string   $addition
-     * @param int|null $location
-     *
-     * @throws \Gnugat\Redaktilo\Search\NotSupportedException If the current line number isn't supported by any registered strategy
+     * @param File    $file
+     * @param string  $addition
+     * @param integer $location
      *
      * @api
      */
-    public function addAfter(File $file, $addition, $location=null)
+    public function addAfter(File $file, $addition, $location = null)
     {
-        if ($location === null || !is_integer($location)) {
-            $location = $file->getCurrentLineNumber();
-        }
-        $location++;
-
         $input = array(
             'file' => $file,
             'location' => $location,
             'addition' => $addition,
         );
-        $this->commandInvoker->run('insert', $input);
-
-        $file->setCurrentLineNumber($location);
+        $this->commandInvoker->run('insert_under', $input);
     }
 
     /**
-     * Changes the current line to the given line.
+     * Replaces the line at the given line number
+     * (or at the current one if none provided) with the given replacement.
      *
-     * @param File     $file
-     * @param string   $replacement
-     * @param int|null $location
-     *
-     * @throws \Gnugat\Redaktilo\Search\NotSupportedException If the current line number isn't supported by any registered strategy
+     * @param File    $file
+     * @param string  $replacement
+     * @param integer $location
      *
      * @api
      */
-    public function changeTo(File $file, $replacement, $location=null)
+    public function changeTo(File $file, $replacement, $location = null)
     {
         if ($location === null || !is_integer($location)) {
             $location = $file->getCurrentLineNumber();
@@ -242,57 +226,20 @@ class Editor
     }
 
     /**
-     * Replaces the current line using a regex and replace string/callback.
+     * Removes the line at the given location
+     * (or at the current one if none provided).
      *
-     * @param File            $file
-     * @param string          $regex
-     * @param string|callable $replace
-     * @param int|null        $location
-     *
-     * @throws \InvalidArgumentException If $replace is not a valid callable or regex
+     * @param File    $file
+     * @param integer $location
      *
      * @api
      */
-    public function replaceWith(File $file, $regex, $replace, $location=null)
+    public function remove(File $file, $location = null)
     {
-        if ($location === null || !is_integer($location)) {
-            $location = $file->getCurrentLineNumber();
-        }
-
-        $converter = new \Gnugat\Redaktilo\Converter\LineContentConverter();
-        $lines = $converter->from($file);
-        $line = $lines[$location];
-
-        if (is_callable($replace)) {
-            $line = preg_replace_callback($regex, $replace, $line);
-        } elseif (is_string($replace)) {
-            $line = preg_replace($regex, $replace, $line);
-        } else {
-            throw new \InvalidArgumentException(sprintf('Expected a callable or valid regex as third argument to Edit#replaceWith(), got "%s".', $replace));
-        }
-        $file->changeLineTo($line, $location);
-
-        $file->setCurrentLineNumber($location);
-    }
-
-    /**
-     * @param File     $file
-     * @param int|null $location
-     *
-     * @api
-     */
-    public function remove(File $file, $location=null)
-    {
-        if ($location === null || !is_integer($location)) {
-            $location = $file->getCurrentLineNumber();
-        }
-
         $input = array(
             'file' => $file,
             'location' => $location,
         );
         $this->commandInvoker->run('remove', $input);
-
-        $file->setCurrentLineNumber($location);
     }
 }