+++ CHANGELOG.md
@@ -1,14 +1,21 @@
 # Changes between versions
 
-## 1.0.0-alpha1: Commands and boolean finds
+## 1.0.0-alpha1: Commands, boolean finds and locations
 
+* removed `SubstringSearchStrategy`
+* removed `replaceWith` from `Editor`
+* added `LineInsertAboveCommand`
+* added `LineInsertUnderCommand`
+* removed `LineInsertCommand`
+* moved line management from `Editor` to commands
+* renamed `UnsupportedCommandException` into `CommandNotFoundException`
 * added before/after to the `Editor` jumpTo methods
 * removed `FactoryMethod`
 * removed `SearchStrategy`'s `has` method
 * added before/after to the `SearchStrategy` find methods
 * added abstract class `LineSearchStrategy`
 * renamed `LineSearchStrategy` to `SameSearchStrategy`
-* moved search excpetion throwing from `Search` to the `Editor`
+* moved search exception throwing from `Search` to the `Editor`
 * added location for `Editor`'s manipulation methods
 * moved `NotSupportedException` from `Engine` to `Search`
 * replaced `ReplaceEngine` with `CommandInvoker`